Add .NET workflow for building, restoring dependencies, and publishing
This commit adds a new file `.github:workflows/dotnet.yml` that contains a workflow for building, restoring dependencies, and publishing a .NET project. The workflow is triggered on push and pull request events on the `main` branch. It runs on an Ubuntu environment and performs the following steps:
- Checks out the repository
- Sets up .NET version 8.0.x
- Restores dependencies using `dotnet restore`
- Builds the project using `dotnet build --configuration Release --no-restore`
- Publishes the project as a NuGet package using brandedoutcast/publish-nuget action with specified PROJECT_FILE_PATH and NUGET_KEY.

Additionally, this commit updates `.gitignore` to exclude `.idea/` folder in `ProcessingPipeline.ImageSharp`.