Updated nuget publish action
The commit updates the version of the NuGet publish action used in the dotnet.yml file from v2.5.5 to v2.5.2. This change ensures that the correct version of the action is being utilized for publishing the project.
Update NuGet publish action version in dotnet.yml

The commit updates the version of the NuGet publish action used in the dotnet.yml file from v2.5.5 to v2.5.2. This change ensures that the correct version of the action is being utilized for publishing the project.